[KYUUBI #1451] Add Row level filtering support
